Split widget files
