+++ packages/dear-imgui/cpp/include/widget/slider.h
@@ -1,0 +1,61 @@
+#include "styled_widget.h"
+
+class Slider final : public StyledWidget {
+    protected:
+        Slider(ReactImgui* view, const int id, const std::string& label, const float defaultValue, const float min, const float max, const std::string& sliderType, std::optional<BaseStyle>& style) : StyledWidget(view, id, style) {
+            m_type = "Slider";
+            m_sliderType = sliderType;
+            m_label = label;
+            m_value = defaultValue;
+            m_min = min;
+            m_max = max;
+        }
+
+    public:
+        std::string m_sliderType;
+        float m_value;
+        float m_min;
+        float m_max;
+        std::string m_label;
+
+        static std::unique_ptr<Slider> makeWidget(const json& widgetDef, std::optional<BaseStyle> maybeStyle, ReactImgui* view) {
+            if (widgetDef.is_object() && widgetDef.contains("id") && widgetDef["id"].is_number_integer()) {
+                const auto id = widgetDef["id"].template get<int>();
+                const auto defaultValue = widgetDef.contains("defaultValue") && widgetDef["defaultValue"].is_number() ? widgetDef["defaultValue"].template get<float>() : 0.0f;
+                const auto min = widgetDef.contains("min") && widgetDef["min"].is_number() ? widgetDef["min"].template get<float>() : 0.0f;
+                const auto max = widgetDef.contains("max") && widgetDef["max"].is_number() ? widgetDef["max"].template get<float>() : 10.0f;
+                const auto label = widgetDef.contains("label") && widgetDef["label"].is_string() ? widgetDef["label"].template get<std::string>() : "";
+                const auto sliderType = widgetDef.contains("sliderType") && widgetDef["sliderType"].is_string() ? widgetDef["sliderType"].template get<std::string>() : "default";
+
+                return makeWidget(view, id, label, defaultValue, min, max, sliderType, maybeStyle);
+            }
+
+            throw std::invalid_argument("Invalid JSON data");
+        }
+
+        static std::unique_ptr<Slider> makeWidget(ReactImgui* view, const int id, const std::string& label, const float defaultValue, const float min, const float max, const std::string& sliderType, std::optional<BaseStyle>& style) {
+            Slider instance(view, id, label, defaultValue, min, max, sliderType, style);
+            return std::make_unique<Slider>(std::move(instance));
+        }
+
+        static YGSize Measure(const YGNodeConstRef node, float width, YGMeasureMode widthMode, float height, YGMeasureMode heightMode) {
+            auto widget = static_cast<Slider*>(YGNodeGetContext(node));
+            YGSize size;
+
+            size.width = 60.0f; // TODO: This is rather arbitrary
+            size.height = widget->m_view->GetFrameHeight(widget);
+
+            return size;
+        }
+
+        void Render(ReactImgui* view) override;
+
+        void Patch(const json& widgetPatchDef, ReactImgui* view) override;
+
+        void Init() override {
+            Element::Init();
+
+            YGNodeSetContext(m_layoutNode->m_node, this);
+            YGNodeSetMeasureFunc(m_layoutNode->m_node, Measure);
+        }
+};
