+++ packages/dear-imgui/cpp/src/widget/checkbox.cpp
@@ -1,0 +1,25 @@
+#include "widget/checkbox.h"
+
+#include <imgui.h>
+
+#include <nlohmann/json.hpp>
+
+#include "reactimgui.h"
+
+void Checkbox::Render(ReactImgui* view) {
+    ImGui::PushID(m_id);
+    if (ImGui::Checkbox(m_label.c_str(), &m_checked)) {
+        view->m_onBooleanValueChange(m_id, m_checked);
+    }
+    ImGui::PopID();
+};
+
+void Checkbox::Patch(const json& widgetPatchDef, ReactImgui* view) {
+    if (widgetPatchDef.is_object()) {
+        StyledWidget::Patch(widgetPatchDef, view);
+
+        if (widgetPatchDef.contains("label") && widgetPatchDef["label"].is_string()) {
+            m_label = widgetPatchDef["label"].template get<std::string>();
+        }
+    }
+};
