+++ packages/dear-imgui/cpp/src/widget/separator.cpp
@@ -1,0 +1,7 @@
+#include "widget/separator.h"
+
+#include "reactimgui.h"
+
+void Separator::Render(ReactImgui* view) {
+    ImGui::Separator();
+};
